[refactor] Remove inference module
