+++ astroid/constraint.py
@@ -8,15 +8,18 @@ from __future__ import annotations
 import sys
 from abc import ABC, abstractmethod
 from collections.abc import Iterator
-from typing import Union
+from typing import TYPE_CHECKING, Union
 
-from astroid import bases, nodes, util
+from astroid import nodes, util
 from astroid.typing import InferenceResult
 
 if sys.version_info >= (3, 11):
     from typing import Self
 else:
     from typing_extensions import Self
+
+if TYPE_CHECKING:
+    from astroid import bases
 
 _NameNodes = Union[nodes.AssignAttr, nodes.Attribute, nodes.AssignName, nodes.Name]
